Rewrite specs to execute all queries
This commit rewrites all of the specs to execute real queries against
the database, and completely tear down all tables after each spec. This
provides better protection against order-dependent specs, breaking MySQL
changes, and will allow our specs to support ActiveRecord 5+.
