+++ spec/support/connection_helpers.rb
@@ -1,0 +1,13 @@
+module ConnectionHelpers
+  def add_polymorphic_constraints(*args)
+    connection.add_polymorphic_constraints(*args)
+  end
+
+  def remove_polymorphic_constraints(*args)
+    connection.remove_polymorphic_constraints(*args)
+  end
+
+  def triggers
+    connection.triggers
+  end
+end
